# services/api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API dla systemu monitorowania IoT.",
        "title": "IoT Monitor API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/skni-kod/iot-monitor-backend"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/alert-rules": {
            "get": {
                "description": "Get a list of alert rules for the authenticated user with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alert Rules"
                ],
                "summary": "List Alert Rules",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of alert rules",
                        "schema": {
                            "$ref": "#/definitions/types.PaginatedAlertRuleResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new alert rule for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alert Rules"
                ],
                "summary": "Create Alert Rule",
                "parameters": [
                    {
                        "description": "Alert Rule Request",
                        "name": "alertRule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AlertRuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.AlertRuleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/alert-rules/{id}": {
            "get": {
                "description": "Get an alert rule by ID",
                "tags": [
                    "Alert Rules"
                ],
                "summary": "Get Alert Rule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Alert Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.AlertRuleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing alert rule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alert Rules"
                ],
                "summary": "Update Alert Rule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Alert Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Alert Rule Request",
                        "name": "updateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateAlertRuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.AlertRuleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an alert rule by ID",
                "tags": [
                    "Alert Rules"
                ],
                "summary": "Delete Alert Rule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Alert Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.AlertRuleRequest": {
            "type": "object",
            "properties": {
                "condition_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sensor_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "number"
                }
            }
        },
        "types.AlertRuleResponse": {
            "type": "object",
            "properties": {
                "condition_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sensor_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "types.PaginatedAlertRuleResponse": {
            "type": "object",
            "properties": {
                "alert_rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.AlertRuleResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "types.UpdateAlertRuleRequest": {
            "type": "object",
            "properties": {
                "condition_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sensor_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Wprowadź token JWT w formacie 'Bearer {token}'.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}